+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added diagnostic for relative humidity from temperature and dew point
 - Added diagnostic for wind speed magnitude
 - Added diagnostic for vapour-pressure deficit
+- Added ForecastInterpolation model for temporal interpolation of forecasts
 
 ### Changed